Documented permissions.
Also reorganized a bit to make documentation easier to navigate.
